requirements.txt
fastapi
uvicorn
python-multipart
Pillow
transformers
torch
accelerate
opencv-python-headless
